moved enum for zoom boundaries
