+++ src/main/java/jfxtras/labs/map/tile/GoogleTileSourceFactory.java
@@ -8,8 +8,6 @@ import javafx.scene.image.Image;
  */
 public class GoogleTileSourceFactory extends ApiKeyTileSourceFactory<GoogleType>{
     
-    private static final int ZOOM = 21;
-    
     private static final String TILE_URL =  "http://%s.google.com/vt";
 
     public GoogleTileSourceFactory(String apiKey) {
@@ -32,7 +30,6 @@ public class GoogleTileSourceFactory extends ApiKeyTileSourceFactory<GoogleType>
         
         TilePathBuildable tilePathBuilder = new GoogleTilePathBuilder();
         tileSource.setTilePathBuilder(tilePathBuilder);
-        tileSource.setMaxZoom(ZOOM);
         
         Image image = new Image(getClass().getResourceAsStream("google_white.png"));
         tileSource.setAttributionImage(image);
