+++ src/main/java/jfxtras/labs/map/tile/TileRepository.java
@@ -62,7 +62,6 @@ public class TileRepository {
             cleanupCache();
 
             String location = tileSource.getTileUrl(zoom, tilex, tiley);
-            System.err.println(location);
             TileInfo info = cache.get(location);
 
             if (info != null) {
